# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Taskosaur API",
    "description": "API for Taskosaur - Task Management Application",
    "version": "1.0.0",
    "contact": {
      "email": "info@taskosaur.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["users"],
        "summary": "Get all users",
        "description": "Returns all users the authenticated user has permission to view",
        "operationId": "getUsers",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          }
        }
      },
      "post": {
        "tags": ["users"],
        "summary": "Create user",
        "description": "Create a new user (Admin only)",
        "operationId": "createUser",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "description": "User to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": ["users"],
        "summary": "Get user by ID",
        "description": "Returns a user by ID",
        "operationId": "getUserById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "tags": ["users"],
        "summary": "Update user",
        "description": "Update an existing user",
        "operationId": "updateUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "User details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Email already in use"
          }
        }
      },
      "delete": {
        "tags": ["users"],
        "summary": "Delete user",
        "description": "Delete a user (Admin only)",
        "operationId": "deleteUser",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
...
    "/workspaces/{workspaceId}/members/{userId}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get workspace member (Detailed)",
        "description": "Get details of a specific workspace member",
        "operationId": "getWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMember"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      },
      "put": {
        "tags": ["roles"],
        "summary": "Update workspace member (Detailed)",
        "description": "Update a workspace member's role",
        "operationId": "updateWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Member details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceMember"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Remove workspace member (Detailed)",
        "description": "Remove a member from a workspace",
        "operationId": "removeWorkspaceMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "description": "ID of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace member removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace member not found"
          }
        }
      }
    },
...
    "/projects/{projectId}/members/{userId}": {
      "get": {
        "tags": ["roles"],
        "summary": "Get project member (Detailed)",
        "description": "Get details of a specific project member",
        "operationId": "getProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project member details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMember"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      },
      "put": {
        "tags": ["roles"],
        "summary": "Update project member (Detailed)",
        "description": "Update a project member's role",
        "operationId": "updateProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Member details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectMemberRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project member updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMember"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      },
      "delete": {
        "tags": ["roles"],
        "summary": "Remove project member (Detailed)",
        "description": "Remove a member from a project",
        "operationId": "removeProjectMemberDetailed",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "ID of the project",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "userId",
            "in": "path",
            "description": "ID of the user",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Project member removed"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Project member not found"
          }
        }
      }
    },
...
    "/workspaces/{slug}": {
      "get": {
        "tags": ["workspaces"],
        "summary": "Get workspace by slug",
        "description": "Returns a single workspace identified by its slug",
        "operationId": "getWorkspaceBySlug",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Workspace details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace not found"
          }
        }
      },
      "put": {
        "tags": ["workspaces"],
        "summary": "Update workspace",
        "description": "Updates an existing workspace",
        "operationId": "updateWorkspace",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Workspace details to update",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Workspace updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace not found"
          },
          "409": {
            "description": "Slug already in use"
          }
        }
      },
      "delete": {
        "tags": ["workspaces"],
        "summary": "Delete workspace",
        "description": "Deletes a workspace",
        "operationId": "deleteWorkspace",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "description": "Slug of the workspace",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Workspace deleted"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - Insufficient permissions"
          },
          "404": {
            "description": "Workspace not found"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "UpdateWorkspaceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Updated Marketing Team"
          },
          "description": {
            "type": "string", 
            "example": "Updated description for the marketing department"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "example": "https://example.com/avatar.png"
          }
        }
      },
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "user-123"
          },
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "role": {
            "type": "string",
            "example": "User",
            "enum": ["Admin", "User"]
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": ["Active", "Inactive", "Pending"]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-10T10:30:00Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          }
        }
      },
      "CreateUserRequest": {
        "type": "object",
        "required": ["name", "email", "password", "role"],
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "example": "password123"
          },
          "role": {
            "type": "string",
            "example": "User",
            "enum": ["Admin", "User"]
          }
        }
      },
      "UpdateUserRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John Doe"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/avatar.png"
          },
          "role": {
            "type": "string",
            "example": "User",
            "enum": ["Admin", "User"]
          },
          "status": {
            "type": "string",
            "example": "Active",
            "enum": ["Active", "Inactive"]
          }
        }
      },
...
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "permission-123"
          },
          "name": {
            "type": "string",
            "example": "workspace:create"
          },
          "description": {
            "type": "string",
            "example": "Permission to create workspaces"
          },
          "resource": {
            "type": "string",
            "example": "workspace"
          },
          "action": {
            "type": "string",
            "example": "create"
          }
        }
      },
      "WorkspaceMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "workspace-member-123"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "workspace": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "workspace-123"
              },
              "name": {
                "type": "string",
                "example": "My Workspace"
              },
              "slug": {
                "type": "string",
                "example": "my-workspace"
              }
            }
          },
          "role": {
            "type": "string",
            "example": "Owner",
            "enum": ["Owner", "Admin", "Member", "Guest"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "UpdateWorkspaceMemberRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "example": "Admin",
            "enum": ["Admin", "Member", "Guest"]
          }
        }
      },
      "ProjectMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "project-member-123"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "project": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "example": "project-123"
              },
              "name": {
                "type": "string",
                "example": "My Project"
              },
              "slug": {
                "type": "string",
                "example": "my-project"
              },
              "workspace": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "example": "workspace-123"
                  },
                  "name": {
                    "type": "string",
                    "example": "My Workspace"
                  },
                  "slug": {
                    "type": "string",
                    "example": "my-workspace"
                  }
                }
              }
            }
          },
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": ["Manager", "Member", "Viewer"]
          },
          "joinedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2023-04-15T14:30:00Z"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
...
      "UpdateProjectMemberRequest": {
        "type": "object",
        "required": ["role"],
        "properties": {
          "role": {
            "type": "string",
            "example": "Manager",
            "enum": ["Manager", "Member", "Viewer"]
          }
        }
      },
...
      "Workspace": {
        "type": "object",
        "required": ["id", "slug", "name", "description", "projectCount", "memberCount", "recentActivity"],
        "properties": {
          "id": {
            "type": "string",
            "example": "ws-abc"
          },
          "slug": {
            "type": "string",
            "example": "my-cool-workspace"
          },
          "name": {
            "type": "string",
            "example": "My Cool Workspace"
          },
          "description": {
            "type": "string",
            "example": "This is a description of the workspace."
          },
          "projectCount": {
            "type": "integer",
            "example": 5
          },
          "memberCount": {
            "type": "integer",
            "example": 12
          },
          "recentActivity": {
            "type": "string",
            "example": "Task 'Deploy to Prod' updated 2 hours ago"
          },
          "avatar": {
            "type": "string",
            "format": "url",
            "nullable": true,
            "example": "https://example.com/workspace-avatar.png"
          }
        }
      },
...
    }
  }
}